+++ Trains3D/Trains3D/Controller/TRCity.m
@@ -92,10 +92,10 @@ static NSArray* _TRCityAngle_values;
 }
 
 - (TRRailPoint*)startPoint {
-    return [TRRailPoint railPointWithTile:_tile form:_angle.form x:0 back:_angle.back];
+    return [TRRailPoint railPointWithTile:_tile form:_angle.form x:((float)(0)) back:_angle.back];
 }
 
-- (void)updateWithDelta:(double)delta {
+- (void)updateWithDelta:(float)delta {
     [_expectedTrainAnimation forEach:^void(EGAnimation* _) {
         [_ updateWithDelta:delta];
     }];
